updated AddressDomainEventPublisherE2ETest to use common assertion logic
