+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/AddressDomainEventPublisherE2ETest.kt
@@ -1,25 +1,18 @@
 package uk.gov.justice.digital.hmpps.personrecord.service
 
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.web.reactive.server.expectBody
-import software.amazon.awssdk.services.sqs.model.ReceiveMessageRequest
-import tools.jackson.module.kotlin.readValue
 import uk.gov.justice.digital.hmpps.personrecord.api.constants.Roles.PROBATION_API_READ_WRITE
 import uk.gov.justice.digital.hmpps.personrecord.api.model.probation.ProbationCreateAddressResponse
-import uk.gov.justice.digital.hmpps.personrecord.client.model.sqs.MessageAttribute
-import uk.gov.justice.digital.hmpps.personrecord.client.model.sqs.SQSMessage
-import uk.gov.justice.digital.hmpps.personrecord.client.model.sqs.messages.domainevent.DomainEvent
 import uk.gov.justice.digital.hmpps.personrecord.config.E2ETestBase
 import uk.gov.justice.digital.hmpps.personrecord.model.person.Address
 import uk.gov.justice.digital.hmpps.personrecord.service.DomainEventSource.CPR
 import uk.gov.justice.digital.hmpps.personrecord.service.address.AddressService
 import uk.gov.justice.digital.hmpps.personrecord.service.type.CPR_PROBATION_ADDRESS_CREATED
 import uk.gov.justice.digital.hmpps.personrecord.test.randomCrn
-import java.util.UUID
 
 class AddressDomainEventPublisherE2ETest : E2ETestBase() {
 
@@ -32,7 +25,7 @@ class AddressDomainEventPublisherE2ETest : E2ETestBase() {
     val newAddress = createRandomProbationAddress()
     createPersonWithNewKey(createRandomProbationPersonDetails(crn).copy(addresses = emptyList()))
 
-    val responseBody = webTestClient
+    webTestClient
       .post()
       .uri(probationAddressApiUrl(crn))
       .headers(jwtAuthorisationHelper.setAuthorisationHeader(roles = listOf(PROBATION_API_READ_WRITE)))
@@ -43,27 +36,10 @@ class AddressDomainEventPublisherE2ETest : E2ETestBase() {
       .expectBody<ProbationCreateAddressResponse>()
       .returnResult().responseBody!!
 
-    val createdAddress = addressRepository.findByUpdateId(UUID.fromString(responseBody.cprAddressId))
-
-    expectOneMessageOn(testOnlyCPRDomainEventsQueue)
-    val rawDomainEventMessage = testOnlyCPRDomainEventsQueue?.sqsClient?.receiveMessage(
-      ReceiveMessageRequest.builder().queueUrl(testOnlyCPRDomainEventsQueue?.queueUrl).build(),
+    assertDomainEventPublishedAfterSasEvent(
+      expectedEventType = CPR_PROBATION_ADDRESS_CREATED,
+      crn = crn,
     )
-    val sqsMessage = rawDomainEventMessage?.get()?.messages()?.first()?.let { jsonMapper.readValue<SQSMessage>(it.body()) }!!
-    assertThat(sqsMessage.messageAttributes?.eventType).isEqualTo(MessageAttribute(CPR_PROBATION_ADDRESS_CREATED))
-    assertThat(sqsMessage.messageAttributes?.eventSource).isEqualTo(MessageAttribute(CPR.identifier))
-    assertThat(sqsMessage.message.contains("unmergedCRN")).isFalse()
-
-    val domainEvent: DomainEvent = jsonMapper.readValue(sqsMessage.message)
-    assertThat(domainEvent.eventType).isEqualTo(CPR_PROBATION_ADDRESS_CREATED)
-    assertThat(domainEvent.detailUrl).isEqualTo("http://localhost:8080/person/probation/$crn/address/${createdAddress?.updateId}")
-    assertThat(domainEvent.description).isEqualTo("A probation address has been created for a person")
-    assertThat(domainEvent.occurredAt).isNotNull()
-    assertThat(domainEvent.personReference?.identifiers?.size).isEqualTo(1)
-    assertThat(domainEvent.personReference?.identifiers?.get(0)?.type).isEqualTo("CRN")
-    assertThat(domainEvent.personReference?.identifiers?.get(0)?.value).isEqualTo(crn)
-    assertThat(domainEvent.additionalInformation?.cprAddressId).isEqualTo(createdAddress?.updateId.toString())
-    assertThat(domainEvent.additionalInformation?.outboundDeliusAddressId).isNull()
   }
 
   @Nested
